fix: fix clearing feedbacks and marks when delete system

+++ feedback-service/src/main/java/org/example/service/FeedbackService.java
@@ -1,6 +1,7 @@
 package org.example.service;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.example.repository.CourseRatingRepository;
 import org.example.repository.ExplorerFeedbackRepository;
 import org.example.repository.KeeperFeedbackRepository;
@@ -10,6 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class FeedbackService {
     private final ExplorerFeedbackRepository explorerFeedbackRepository;
     private final KeeperFeedbackRepository keeperFeedbackRepository;
@@ -18,8 +20,11 @@ public class FeedbackService {
     @KafkaListener(topics = "deleteFeedbackTopic", containerFactory = "deleteFeedbackKafkaListenerContainerFactory")
     @Transactional
     public void deleteFeedbackByExplorerId(Integer explorerId) {
-        explorerFeedbackRepository.deleteById(explorerId);
-        keeperFeedbackRepository.deleteById(explorerId);
-        courseRatingRepository.deleteById(explorerId);
+        if (explorerFeedbackRepository.existsById(explorerId))
+            explorerFeedbackRepository.deleteById(explorerId);
+        if (keeperFeedbackRepository.existsById(explorerId))
+            keeperFeedbackRepository.deleteById(explorerId);
+        if (courseRatingRepository.existsById(explorerId))
+            courseRatingRepository.deleteById(explorerId);
     }
 }
